+++ docs/changes.rst
@@ -7,6 +7,7 @@ Release history
 - feat: new API :attr:`api.scim.groups <wxc_sdk.scim.ScimV2Api.groups>`
 - feat: convergedRecordings support for webhooks
 - feat: new API :attr:`api.converged_recordings <wxc_sdk.WebexSimpleApi.converged_recordings>`
+- feat: new API :attr:`api.telephony.organisation_access_codes <wxc_sdk.telephony.TelephonyApi.organisation_access_codes>`
 
 1.19.0
 ------